[Refactor] room-types.h を廃止してroom-types.h に片寄せした
