+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -4484,9 +4484,6 @@
     <ClInclude Include="..\..\src\player\player-status-flags.h">
       <Filter>player</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\room\room-types.h">
-      <Filter>room</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\room\lake-types.h">
       <Filter>room</Filter>
     </ClInclude>
